+++ lib/api.dart
@@ -15,11 +15,12 @@ export 'package:pinenacl/crypto/hmac_sha512.dart';
 export 'package:pinenacl/crypto/sha_256.dart';
 export 'package:pinenacl/crypto/tweetnacl.dart';
 export 'package:pinenacl/encoding.dart';
+export 'package:pinenacl/hashing.dart';
 export 'package:pinenacl/utils.dart';
 
 part 'src/api/base.dart';
 part 'src/api/key_encryption.dart';
 part 'src/api/signatures.dart';
 
-part 'src/registry/algorythms.dart';
+part 'src/registry/algorithms.dart';
 part 'src/registry/registrar.dart';