+++ lib/crypto/tweetnacl_ext.dart
@@ -7,7 +7,7 @@ class TweetNaClExt {
   static const hmacBytes = 64;
 
   // Extended secret key's length
-  static const int extendedSecretKeyLength = 32;
+  static const int extendedSecretKeyLength = 64;
 
   static int _crypto_auth(Uint8List out, Uint8List m, Uint8List k) {
     HmacSha512.mac(out, m, k);
@@ -37,7 +37,7 @@ class TweetNaClExt {
     return TweetNaCl._vn(x, xoff, y, yoff, 64);
   }
 
-  static int scalar_base(Uint8List pk, Uint8List sk) {
+  static Uint8List scalar_base(Uint8List pk, Uint8List sk) {
     List<Int64List> p = List<Int64List>(4);
 
     p[0] = Int64List(16);
@@ -48,7 +48,7 @@ class TweetNaClExt {
     TweetNaCl._scalarbase(p, sk, 0);
     TweetNaCl._pack(pk, p);
 
-    return 0;
+    return pk;
   }
 
   static Uint8List scalar_base_seed(Uint8List pk, Uint8List seed) {