+++ lib/src/api/signatures.dart
@@ -1,17 +1,5 @@
 part of pinenacl.api;
 
-/*abstract class Sign {
-  Verify get verifyKey;
-  EncryptionMessage sign(List<int> message);
-}
-
-
-abstract class Verify extends ByteList {
-  bool verify({SignatureBase signature, List<int> message});
-  bool verifySignedMessage({EncryptionMessage signedMessage});
-}
-*/
-
 abstract class SignatureBase extends ByteList {}
 
 abstract class EncryptionMessage {
@@ -19,7 +7,7 @@ abstract class EncryptionMessage {
   ByteList get message;
 }
 
-mixin Verify<T extends AlgorythmParams> on ByteList {
+mixin Verify<T extends AlgorithmParams> on ByteList {
   bool verify({SignatureBase signature, List<int> message}) {
     if (signature != null) {
       if (signature.length != TweetNaCl.signatureLength) {
@@ -36,7 +24,7 @@ mixin Verify<T extends AlgorythmParams> on ByteList {
     Uint8List m = Uint8List(message.length);
 
     /// get the signing algorytm
-    AlgorythmParams alg = Registrar.getInstance(T.toString());
+    AlgorithmParams alg = Registrar.getInstance(T.toString());
 
     final result = alg.sigParams
         .verifyAlg(m, -1, Uint8List.fromList(message), 0, message.length, this);
@@ -51,12 +39,12 @@ mixin Verify<T extends AlgorythmParams> on ByteList {
       signature: signedMessage.signature, message: signedMessage.message);
 }
 
-mixin Sign<T extends AlgorythmParams> on ByteList {
+mixin Sign<T extends AlgorithmParams> on ByteList {
   SignedMessage sign(List<int> message) {
     // signed message
     Uint8List sm = Uint8List(message.length + TweetNaCl.signatureLength);
 
-    AlgorythmParams alg = Registrar.getInstance(T.toString());
+    AlgorithmParams alg = Registrar.getInstance(T.toString());
 
     final result = alg.sigParams
         .signAlg(sm, -1, Uint8List.fromList(message), 0, message.length, this);
@@ -64,20 +52,22 @@ mixin Sign<T extends AlgorythmParams> on ByteList {
       throw Exception('Signing the massage is failed');
     }
 
-    return SignedMessage.fromList(signedMessage: sm);
+    return SignedMessage<T>.fromList(signedMessage: sm);
   }
 }
 
-class Signature<T extends AlgorythmParams> extends ByteList
+class Signature<T extends AlgorithmParams> extends ByteList
     implements SignatureBase {
   Signature(List<int> bytes) : super(bytes, bytesLength);
   static const bytesLength = TweetNaCl.signatureLength;
 
   @override
-  Codec get encoder => Registrar.getInstance(T.toString()).signParams.codec;
+  Codec get encoder => Registrar.getInstance(T.toString()).sigParams.codec;
 }
 
-class SignedMessage extends ByteList with Suffix implements EncryptionMessage {
+class SignedMessage<T extends AlgorithmParams> extends ByteList
+    with Suffix
+    implements EncryptionMessage {
   SignedMessage({SignatureBase signature, List<int> message})
       : super(signature + message, signatureLength);
   SignedMessage.fromList({List<int> signedMessage}) : super(signedMessage);
@@ -88,7 +78,7 @@ class SignedMessage extends ByteList with Suffix implements EncryptionMessage {
   static const signatureLength = 64;
 
   @override
-  SignatureBase get signature => Signature(prefix);
+  SignatureBase get signature => Signature<T>(prefix);
 
   @override
   ByteList get message => suffix;