+++ lib/src/registry/algorithms.dart
@@ -1,7 +1,7 @@
 part of pinenacl.api;
 
-/// Algorythm's params of the ECDH
-class Curve25519 extends AlgorythmParams {
+/// Algorithm's params of the ECDH
+class Curve25519 extends AlgorithmParams {
   @override
   PrivateParams get prvParams => const PrivateParams(
       pubAlg: TweetNaCl.crypto_scalarmult_base,
@@ -18,9 +18,15 @@ class Curve25519 extends AlgorythmParams {
   @override
   SignatureParams get sigParams =>
       throw Exception('Signing is not supported on Curve25519\'s ECDH');
+
+  @override
+  Uint8List normalizeBytes(Uint8List k) => k;
+
+  @override
+  bool validateBytes(Uint8List k) => true;
 }
 
-class Ed25519 extends AlgorythmParams {
+class Ed25519 extends AlgorithmParams {
   @override
   PrivateParams get prvParams => const PrivateParams(
       pubAlg: TweetNaClExt.scalar_base_seed,
@@ -39,4 +45,10 @@ class Ed25519 extends AlgorythmParams {
       verifyAlg: TweetNaCl.crypto_sign_open,
       codec: Bech32Encoder(hrp: 'ed25519_sig'),
       length: TweetNaCl.signatureLength);
+
+  @override
+  Uint8List normalizeBytes(Uint8List k) => k;
+
+  @override
+  bool validateBytes(Uint8List k) => true;
 }