+++ CHANGELOG.md
@@ -64,3 +64,6 @@ and `crypto_auth_verify` functions of the `NaCl` library (does not exists in Twe
 - Added API class diagrams.
 - Swapped `Bech32` back to the latest and working pub package.
 - Refactored the `decode` factories.
+
+## 0.1.3-dev.1
+- Added constructor for `EncryptedMessage` class, see ilap/pinenacl-dart#3